Moved headers to `include/` and fixed `CMakeLists.txt` to use globs instead of absolute file references.
