+++ SLAM/include/CameraFrame.h
@@ -1,0 +1,18 @@
+//
+// Created by Kyler Jensen on 2/15/17.
+//
+
+#ifndef SLAM_CAMERAFRAME_H
+#define SLAM_CAMERAFRAME_H
+
+#include "_extras.h"
+
+struct CameraFrame {
+
+    const Chrono timestamp;
+
+    //TODO
+
+};
+
+#endif //SLAM_CAMERAFRAME_H
