+++ SLAM/include/Pose.h
@@ -5,10 +5,6 @@
 #ifndef SLAM_POSE_H
 #define SLAM_POSE_H
 
-#include "../slam.h"
-
-using namespace slam;
-
 struct Pose {
     //TODO
 };
