--- SLAM/src/camera/ICameraFrameGenerator.h
@@ -1,17 +1,0 @@
-//
-// Created by Kyler Jensen on 2/16/17.
-//
-
-#ifndef SLAM_ICAMERAFRAMEGENERATOR_H
-#define SLAM_ICAMERAFRAMEGENERATOR_H
-
-#include "../slam.h"
-
-using namespace slam;
-
-class ICameraFrameGenerator {
-
-};
-
-
-#endif //SLAM_ICAMERAFRAMEGENERATOR_H
