--- SLAM/src/depth/IDepthMapGenerator.h
@@ -1,17 +1,0 @@
-//
-// Created by Kyler Jensen on 2/16/17.
-//
-
-#ifndef SLAM_IDEPTHMAPGENERATOR_H
-#define SLAM_IDEPTHMAPGENERATOR_H
-
-#include "../slam.h"
-
-using namespace slam;
-
-class IDepthMapGenerator {
-
-};
-
-
-#endif //SLAM_IDEPTHMAPGENERATOR_H
