--- SLAM/src/pose/IPoseGenerator.h
@@ -1,14 +1,0 @@
-//
-// Created by Kyler Jensen on 2/16/17.
-//
-
-#ifndef SLAM_IPOSEGENERATOR_H
-#define SLAM_IPOSEGENERATOR_H
-
-
-class IPoseGenerator {
-
-};
-
-
-#endif //SLAM_IPOSEGENERATOR_H
