--- SLAM/src/slam.h
@@ -1,16 +1,0 @@
-//
-// Created by Kyler Jensen on 2/15/17.
-//
-
-#ifndef SLAM_SLAM_H
-#define SLAM_SLAM_H
-
-#include <chrono>
-
-namespace slam {
-
-    using Timestamp = std::chrono::time_point<std::chrono::steady_clock>;
-
-};
-
-#endif //SLAM_SLAM_H
