+++ SLAM/test/test_main.cpp
@@ -3,4 +3,4 @@
 //
 
 #define CATCH_CONFIG_MAIN
-#include "../include/catch.hpp"
+#include "catch.hpp"
